backend/requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0      
certifi==2022.12.7 
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2
django-allauth==0.54.0
django-colorfield==0.9.0
django-cors-headers==3.13.0
django-database-translation==1.1.5
django-filter==23.2
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==5.0.4
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.2.0
iniconfig==1.1.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==22.0
Pillow==9.3.0
pluggy==0.13.1
psycopg2-binary==2.9.7
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
toml==0.10.2
typing_extensions==4.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.13
zipp==3.11.0